core: fix msvc warning with warn_level=2

+++ libucm/flags.c
@@ -1,5 +1,5 @@
 #include "flags.h"
-#include <inttypes.h>
+#include <stddef.h>
 
 typedef struct
 {
@@ -9,25 +9,25 @@ typedef struct
 static flags_t flags = { 0 };
 
 int
-get_system_flag(unsigned code)
+get_system_flag(uint64_t code)
 {
-    return (flags.v1 & (1 << code));
+    return (flags.v1 & ((uint64_t)1 << code));
 }
 
 void
-set_system_flag(unsigned code)
+set_system_flag(uint64_t code)
 {
-    flags.v1 |= (1 << code);
+    flags.v1 |= ((uint64_t)1 << code);
 }
 
 void
-unset_system_flag(unsigned code)
+unset_system_flag(uint64_t code)
 {
-    flags.v1 &= ~(1 << code);
+    flags.v1 &= ~((uint64_t)1 << code);
 }
 
 int
-unget_system_flag(unsigned code)
+unget_system_flag(uint64_t code)
 {
     int ret = get_system_flag(code);
     unset_system_flag(code);
