+++ libucm/include/flags.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <inttypes.h>
+
 int
-get_system_flag(unsigned code);
+get_system_flag(uint64_t code);
 void
-set_system_flag(unsigned code);
+set_system_flag(uint64_t code);
 void
-unset_system_flag(unsigned code);
+unset_system_flag(uint64_t code);
